+++ apps/indexer/config/config.exs
@@ -66,7 +66,11 @@ config :indexer, Indexer.Fetcher.TokenBalanceOnDemand, threshold: token_balance_
 # config :indexer, Indexer.Fetcher.ReplacedTransaction.Supervisor, disabled?: true
 if System.get_env("POS_STAKING_CONTRACT") do
   config :indexer, Indexer.Fetcher.BlockReward.Supervisor, disabled?: true
+else
+  config :indexer, Indexer.Fetcher.BlockReward.Supervisor, disabled?: false
 end
+
+config :indexer, Indexer.Fetcher.InternalTransaction.Supervisor, disabled?: false
 
 config :indexer, Indexer.Supervisor, enabled: System.get_env("DISABLE_INDEXER") != "true"